Fix adding partition with automatic type and start sector (#1814805)
When adding a fourth partition blivet automatically creates an
extended partition and the new partition is created as logical
inside it. When manually specifying the start sector of the
partition this always fail, because after creating the extended
partition the start is not in the free space inside the extended
partition.
This change sets the desired start sector for the extended
partition and creates the logical partition inside it without
specifying the start sector.