Fix ignoring disk devices with parents or children
For disk-like devices like multipath we should allow to ignore
these by simply ignoring the mpath device or by ignoring all of its
drives.

- when ignoring the "mpatha" device we should also ignore "sda" and
"sdb"
- when ignoring both "sda" and "sdb" we should also ignore "mpatha"
- when ignoring only "sda" we should not ignore "mpatha" (we don't
want to deal with an "incomplete" multipath device in the tree)

This is consistent with the existing behaviour when using exclusive
disks (or "ignoredisks --only-use" in kickstart).

Resolves: rhbz#1866243