Allow specifying 'mode' for the sdist command
and do not run translation canary when creating archive for daily
builds and/or testing.
This also allows building the source archive without pocketlint.
The PO files should be already sufficiently tested by our CI and
the translation platform so there is a low risk of producing
a broken source tarball.