Add support for creating LVM VDO pools and LVM VDO volumes
The pool and the volume are created by one call but these can have
different properties (like size) and are in fact two block devices
when created, we also need to create two devices and add them to
the devicetree. The pool device must be always created first and
the _create function for the VDO volume is a no-op.