Make sure LV name is unique when adding it in device factory
This fixes issue when "moving" new LV in Anaconda to an existing
volume group. When we add the LV for the first time to a newly
created VG we make sure its name is unique but this can change
when moving the LV to another VG where an LV with same name may
exist. This isn't really a bug because Anaconda tells us to use
the non-unique name but it makes using device factory in this case
more user friendly.
This fixes only LVM because we force system-wide unique names for
other device types so this issue doesn't exist when moving a btrfs
subvolume to another volume.

Resolves: rhbz#1831373