Fix external dependencies for LVM VDO devices
The external and unavailable dependencies code is mostly supposed
to work with just class objects and not instances, which is problem
for LVM devices where the LVMLogicalVolumeDevice can't depend on
LVM VDO and special LVM VDO device mixin classes don't inherit
from the Device class so they are missing some availability
functions.
This fix adds the neccessary functions to LVM VDO mixin classes to
make sure both "unavailable_type_dependencies" and
"type_external_dependencies" work with LVMVDOLogicalVolumeMixin
and LVMVDOPoolMixin. When working with an LVMLogicalVolumeDevice
instance its dependencies are correctly set based on type of the
logical volume.