Do not log entire exception when trying to get ISCSI initiator name
We do this every time and in most cases the call will fail
because there is no ISCSI firmware. Having the entire exception
logged is not useful for us (it will always be the same backtrace
for a failed DBus call) and it confuses users when reporting bugs
because it's the first error message in the log.