Add support for working with locked Stratis pools
Locked pools are tricky because Stratis uses LUKS but the "Stratis
LUKS devices" can't be unlocked with cryptsetup, only with Stratis
API. This implementation "hides" the LUKS device for locked pools
and use the StratisBlockdev format for them instead. To unlock
a locked pool, unlock_pool() function of the StratisBlockdev format
must be used.