Switch LVM devices filter from "reject" to "accept" by default
We currently use the LVM reject filter to filter out hidden and
ignored devices, this commit changes the behaviour to reject all
devices by default and accept only physical volumes that are not
hidden or ignored. This is preparation for the switch from the
existing lvm.conf based filtering to the new devices file based
filtering introduced in LVM 2.03.12 which allows only listing
"accepted" devices. This allows us to support both the "old" and
"new" style filtering using the same code.