Use subvolume mountpoints when listing btrfs subvolumes
We are currently using the btrfs volume mountpoint when trying to
get list of subvolumes and if the volume isn't mounted, we don't
show the subvolumes. But if one of the subvolumes is mounted
(like with the default Fedora installation where / and /home are
mounted) we can also use the subvolume mountpoint to run the
"btrfs subvol list <mnt>" command. This change makes sure we check
all the possible mountpoints to show the subvolumes.

Resolves: rhbz#1856048