Split the test suite into "unit" and "storage" tests
The idea is the unit tests don't actually use the storage, we are
mostly using Mock and patch to test the internal logic of Blivet
and these tests can be run without root privileges. On the other
hand the storage tests use "real" storage devices (loop devices
or disk images), need root privileges and can potentially damage
the system.