Assure that tests that set LVM devices filter can run without root
The blivet.devicelibs.lvm needs_config_refresh decorator checks
whether the libblockdev LVM plugin is available and this check
will fail when not running as root (because LVM DBus API can be
called only by a privileged user). Unfortunately functions that
are decorated by this decorator are replaced by the no-op function
at import time so we can't patch the availability check when the
test is being executed and we need to patch the lvm_devices_*
functions instead.