Do not check for "problematic" disks in ActionList._pre_process
The _find_active_devices_on_action_disks function originally
prevented from making any changes on disks with active devices
(active LVs, mounted partitions etc.) This was changed in
b72e957 and the check currently prevents only reformatting
the disklabel on such disks which should be already impossible
on disks with an existing partition.

The check currently also makes it impossible to add a new disk to
an existing VG -- the new disk has an active device (the VG) in
our storage model so creating a new partition table (even if it
didn't have one before) marks it as "problematic".