Add a forced delay to udev settle in chroot environments
When running in a chroot environment (e.g.mock) there is a race between
udev actually settling the discovered loopback/dm device names and a
program running 'udevadm settle', as latter becomes a no-op in chroot:

   <mock-chroot> sh-5.1# udevadm settle
   Running in chroot, ignoring request.

This patch tries to determine whether it is running in chroot, mimicking
the actions of systemd's running_in_chroot(), and do a 1s delay instead
of settling with udevadm.

1s delay was chosen arbitrary, to succeed in used environment.
However, it should be sufficient for most chrooted environments, as
significant device changes are not expected there.