devices/partition: add ability to auto apply a GPT UUID
Add optional parameter 'mountpoint' for the PartitionDevice
object constructor. This is the path that the partition is
intended to be mounted on. Note this does not imply the
partition will be directly formatted with a filesystem,
there may be intermediate formats such LUKS layered in
before the filesystem.

If the parent disk uses the 'gpt' disk label, then the
mountpoint is used to lookup the well known partition
type UUID associated with that path. This enables the
OS to automatically discover which partitions can be
used without needing to have an fstab.

The partition auto-discovery logic is not enabled by
default for the sake of backwards compatibility. It
usage requires an opt-in by setting a flag

  from blivet.flags import flags
  flags.gpt_discoverable_partitions = True

This means that code can be changed to unconditionally
pass 'mountpoint' into PartitionDevice wherever it is
available. The impact of passing this will be dormant
until the flag is set to enable its usage.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>