Check physical and logical block size when creating a LUKS format
For disks with 4k physical sector size and 512 logical sector size
cryptsetup creates a 4k dm-crypt devicei. This is a problem when
creating LVM on top of two 512 logical sector size disks where
we end up with two incompatible devices. So in this case we need
to explicitly tell cryptsetup to use the logical sector size
instead of letting it decide the optimal sector size for the
dm-crypt device.