Always prefer GPT disk labels on x86_64 (and clean up the logic)
See: https://bugzilla.redhat.com/show_bug.cgi?id=2092091#c6
There was a Fedora 37 Change to prefer GPT disk labels on x86_64
BIOS installs, but for some reason, this was implemented in
anaconda by having it ignore blivet's ordering of the disk label
types and just universally prefer GPT if it's in the list at all,
which resulted in a preference for GPT in more cases than just
x86_64 BIOS. This is one step towards fixing that, by putting the
'always prefer GPT on x86_64' logic here in the blivet ordering
code where it belongs. Step 2 will be to drop the anaconda code
that overrides blivet's preference order.

This also simplifies the logic a bit; it had gotten rather a lot
of conditions piled on top of each other and was rather hard to
read. This should achieve the same effect as before in a clearer
and more concise way.

Signed-off-by: Adam Williamson <awilliam@redhat.com>