Include btrfs volumes names/labels in DeviceTreeBase.names
Btrfs Volumes use label of the filesystem as name in blivet, which
isn't really a name, so it doesn't have to be unique, but not
having it unique causes problems in both Blivet and Anaconda where
we assume device name is a unique identifier. Not having the name
in the list of names means newly created devices can get the same
name as an already existing (or scheduled to be created) device.