Added missing fstab object to SwapSpace
Anaconda tests revealed missing FSTabOptions object in SwapSpace class.
This issue is related to the recent merge of fstab rework PR #1119 which
added this object to FS class. SwapSpace does not inherit from FS and
was overlooked. This change adds the missing object.

It also adds logic that skips all fstab related operations if fstab file
is not to be written (i.e. fstab.dest_file is None). The reason for this is to
eliminate the risk of potential issues caused by unused component.

The test for above has been added as well.