Add "device ID" that could be used as a unique device identifier
Currently device name is commonly used as a unique device
identifier both in blivet internally and in Anaconda. The problem
with this is that duplicate names are allowed in the system in
some cases -- btrfs names doesn't need to be unique and we can
also have two different devices with the same name if they use
different technologies (e.g. an LVM VG and MD array can share the
same name). The new "device_ID" parameter is designed to be usable
as a unique identifier and to be different for devices that share
the same name.

The device ID looks as follows:
- name for disks, partitions and loop devices
- LVM-<name> for LVM VGs and LVs
- BTRFS-<uuid>-<name> for btrfs volumes and subvolumes
- STRATIS-<name> for Stratis devices
- LUKS-<name> for LUKS mapped devices
- MDRAID-<name> for MD arrays
- DM-<name> for generic DM devices