Remove support for NVDIMM namespaces
NVDIMM drives never really became a thing and are no longer
produced. Anaconda will remove support for NVDIMM devices and
the libblockdev NVDIMM plugin is now deprecated. NVDIMM namespaces
that are configured in the block/storage mode will still be usable
as generic DiskDevices, we are removing the NVDIMM specific code
and ability to reconfigure the namespaces.