free_space_estimate: adjust for compression on btrfs
We have a problem with Fedora aarch64 KDE disk images ATM. We
want to make them 16 power-of-ten gigabytes big, so they will fit
on "16GB" SD cards / USB sticks. When we set them to this size,
compose fails because the anaconda free space check fails, saying
"An additional 436 MiB is needed". That check is ultimately backed
by this estimate.

In these images, the main system partition is a btrfs volume,
with compression enabled. If you examine the last successful
image (which was built at size 18 power-of-two GB), thanks to
compression, only ~5GB of space is actually occupied on that
volume. The contents would easily fit if it were slightly smaller.

With this change we adjust the free space estimate by a
compression factor, as well as a metadata factor, if the class
defines one. For the btrfs class we set it to 1.4 when the
compression flag is set. That's a pretty conservative estimate,
we will usually achieve a much better compression ratio with
a typical OS payload (as the numbers above show). AFAICT this
codepath is only actually used by the anaconda size check, so
the change should be fairly safe.

Signed-off-by: Adam Williamson <awilliam@redhat.com>