Fix handling devices with "no" parents in udev
Another attempt to make blivet.reset() more stable and not crash
when encountering broken devices. This is triggered by lingering
MD device after removing it's parent devices. The MD device itself
will be "broken" and unusable in blivet, but at least the reset/
populate call won't end with an error.