Drop parted device cache during reset
When recreating an MD RAID device with different size outside
blivet, during our second device scan the parted device we get
from parted reports the size of the original MD array. This causes
parted to think the GPT partition table on the device has a wrong
size, offers to fix it and destroying it in the process.

Resolves: rhbz#2357214