Fix getting missing libblockdev technologies with Python 3.14
Hotfix for rhbz#2371256. Technologies in libblockdev are defined
as enums and we are using 'value_name' to get the technology name/
description in blivet. For some unknown reason the 'value_name'
attribute from GLib.GEnumValue is not available with Python 3.14.
This needs to be investigated more thoroughly, but for now we can
just not include this information.

Resolves: rhbz#2371256