Group DEVICE_TYPE_* constants in an Enum
For other code using blivet, it is useful to know all of the device types that blivet supports.
This is better accomplished by grouping all of the device type constants into an Enum type so that
the consuming code can import just the Enum and then retrieve all the values it needs from there.

This code makes the Enum the canonical source for that information but also recreates the variables
that have existed up to now from the new Enum.