Tell LVM DBus to refresh it's internal status during reset
Unfortunately some users run wipefs <disk> thinking it's enough
to remove all devices on top of the disk cleanly.
In cases where the PV is not directly on the disk, LVM DBus
doesn't get a udev event and doesn't remove the VG and LVs from
DBus so we think these still exist.

Resolves: RHEL-93967